# Add span-based Uri {Try}{Un}EscapeDataString overloads