--- a/src/libraries/System.Private.Uri/src/System/UriHelper.cs
+++ b/src/libraries/System.Private.Uri/src/System/UriHelper.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Buffers;
@@ -118,26 +118,79 @@ namespace System
             return true;
         }
 
-        internal static string EscapeString(string stringToEscape, bool checkExistingEscaped, SearchValues<char> noEscape)
+        public static bool TryEscapeDataString(ReadOnlySpan<char> charsToEscape, Span<char> destination, out int charsWritten)
+        {
+            if (destination.Length < charsToEscape.Length)
+            {
+                charsWritten = 0;
+                return false;
+            }
+
+            int indexOfFirstToEscape = charsToEscape.IndexOfAnyExcept(Unreserved);
+            if (indexOfFirstToEscape < 0)
+            {
+                // Nothing to escape, just copy the original chars.
+                charsToEscape.CopyTo(destination);
+                charsWritten = charsToEscape.Length;
+                return true;
+            }
+
+            var vsb = new ValueStringBuilder(stackalloc char[Uri.StackallocThreshold]);
+
+            // This may throw for very large inputs. ¯\_(ツ)_/¯
+            vsb.EnsureCapacity(charsToEscape.Length - indexOfFirstToEscape);
+
+            EscapeStringToBuilder(charsToEscape.Slice(indexOfFirstToEscape), ref vsb, Unreserved, checkExistingEscaped: false);
+
+            int newLength = checked(indexOfFirstToEscape + vsb.Length);
+            Debug.Assert(newLength > charsToEscape.Length);
+
+            if (destination.Length >= newLength)
+            {
+                charsToEscape.Slice(0, indexOfFirstToEscape).CopyTo(destination);
+                vsb.AsSpan().CopyTo(destination.Slice(indexOfFirstToEscape));
+
+                vsb.Dispose();
+                charsWritten = newLength;
+                return true;
+            }
+
+            vsb.Dispose();
+            charsWritten = 0;
+            return false;
+        }
+
+        public static string EscapeString(string stringToEscape, bool checkExistingEscaped, SearchValues<char> noEscape)
         {
             ArgumentNullException.ThrowIfNull(stringToEscape);
 
-            Debug.Assert(!noEscape.Contains('%'), "Need to treat % specially; it should be part of any escaped set");
+            return EscapeString(stringToEscape, checkExistingEscaped, noEscape, stringToEscape);
+        }
 
-            int indexOfFirstToEscape = stringToEscape.AsSpan().IndexOfAnyExcept(noEscape);
+        public static string EscapeString(ReadOnlySpan<char> charsToEscape, bool checkExistingEscaped, SearchValues<char> noEscape, string? backingString)
+        {
+            Debug.Assert(!noEscape.Contains('%'), "Need to treat % specially; it should be part of any escaped set");
+            Debug.Assert(backingString is null || backingString.Length == charsToEscape.Length);
+
+            int indexOfFirstToEscape = charsToEscape.IndexOfAnyExcept(noEscape);
             if (indexOfFirstToEscape < 0)
             {
-                // Nothing to escape, just return the original string.
-                return stringToEscape;
+                // Nothing to escape, just return the original value.
+                return backingString ?? charsToEscape.ToString();
             }
 
             // Otherwise, create a ValueStringBuilder to store the escaped data into,
-            // append to it all of the noEscape chars we already iterated through,
-            // escape the rest, and return the result as a string.
+            // escape the rest, and concat the result with the characters we skipped above.
             var vsb = new ValueStringBuilder(stackalloc char[Uri.StackallocThreshold]);
-            vsb.Append(stringToEscape.AsSpan(0, indexOfFirstToEscape));
-            EscapeStringToBuilder(stringToEscape.AsSpan(indexOfFirstToEscape), ref vsb, noEscape, checkExistingEscaped);
-            return vsb.ToString();
+
+            // This may throw for very large inputs. ¯\_(ツ)_/¯
+            vsb.EnsureCapacity(charsToEscape.Length - indexOfFirstToEscape);
+
+            EscapeStringToBuilder(charsToEscape.Slice(indexOfFirstToEscape), ref vsb, noEscape, checkExistingEscaped);
+
+            string result = string.Concat(charsToEscape.Slice(0, indexOfFirstToEscape), vsb.AsSpan());
+            vsb.Dispose();
+            return result;
         }
 
         internal static unsafe void EscapeString(scoped ReadOnlySpan<char> stringToEscape, ref ValueStringBuilder dest,