… src/libraries/System.Private.Uri/tests/FunctionalTests/PercentEncodingHelperTests.cs
… src/libraries/System.Private.Uri/tests/FunctionalTests/UriEscapingTest.cs
… src/libraries/System.Runtime/tests/System.Runtime.Tests/System/Uri.MethodsTests.cs