Closes #40603

```c#
namespace System
{
    partial class Uri
    {
        public static string EscapeDataString(ReadOnlySpan<char> charsToEscape);
        public static bool TryEscapeDataString(ReadOnlySpan<char> charsToEscape, Span<char> destination, out int charsWritten);

        public static string UnescapeDataString(ReadOnlySpan<char> charsToUnescape);
        public static bool TryUnescapeDataString(ReadOnlySpan<char> charsToUnescape, Span<char> destination, out int charsWritten);
    }
}
```

The test coverage for these was not great, so I added a bunch more combinations. I also moved similar test cases from `System.Runtime.Tests` into the Uri project.

I updated `FormUrlEncodedContent` to also use these, so it's now allocating a lot less. For a random usage I pulled from [eShopOnWeb tests](https://github.com/dotnet-architecture/eShopOnWeb/blob/2fa4a3d3368381818c6191122d037c6a70b302f5/tests/FunctionalTests/Web/Controllers/AccountControllerSignIn.cs#L63-L69):
| Method             | Toolchain  | Mean     | Error   | Ratio | Allocated | Alloc Ratio |
|------------------- |----------- |---------:|--------:|------:|----------:|------------:|
| EShopOnWebLogin    | main       | 513.7 ns | 1.69 ns |  1.00 |    2080 B |        1.00 |
| EShopOnWebLogin    | pr         | 291.3 ns | 1.47 ns |  0.57 |     592 B |        0.28 |

Looking at something like YARP's [`EncodePath`](https://github.com/microsoft/reverse-proxy/blob/379671ef2afd01e55168929c33c6b5cbcd0b06b0/src/ReverseProxy/Forwarder/RequestUtilities.cs#L132) helper with ASP.NET Core's [test paths](https://github.com/dotnet/aspnetcore/blob/14868cc8741fda987431003ea3b06ac01e7fc3f1/src/Http/Http.Abstractions/perf/Microbenchmarks/PathStringBenchmark.cs#L10-L12), the difference is
| Method | Scenario             | Mean      | Error    | Ratio | Allocated | Alloc Ratio |
|------- |--------------------- |----------:|---------:|------:|----------:|------------:|
| Old    | LongTestPath         |  95.51 ns | 0.702 ns |  1.00 |     400 B |        1.00 |
| New    | LongTestPath         |  70.58 ns | 0.291 ns |  0.74 |     344 B |        0.86 |
|        |                      |           |          |       |           |             |
| Old    | LongT(...)rcent [24] | 337.00 ns | 2.713 ns |  1.00 |     392 B |        1.00 |
| New    | LongT(...)rcent [24] | 198.49 ns | 0.892 ns |  0.59 |     280 B |        0.71 |
|        |                      |           |          |       |           |             |
| Old    | TestPath             |  67.28 ns | 0.439 ns |  1.00 |     112 B |        1.00 |
| New    | TestPath             |  44.98 ns | 0.129 ns |  0.67 |      56 B |        0.50 |